split examples and pytests
