+++ tests/tests_app/utilities/test_load_app.py
@@ -3,17 +3,12 @@ from unittest.mock import ANY
 
 import pytest
 import tests_app.core.scripts
-from tests_app import _PROJECT_ROOT
 
-from lightning_app import LightningApp
 from lightning_app.utilities.exceptions import MisconfigurationException
 from lightning_app.utilities.load_app import extract_metadata_from_app, load_app_from_file
 
 
 def test_load_app_from_file():
-    app = load_app_from_file(os.path.join(_PROJECT_ROOT, "examples", "app_v0", "app.py"))
-    assert isinstance(app, LightningApp)
-
     test_script_dir = os.path.join(os.path.dirname(tests_app.core.__file__), "scripts")
     with pytest.raises(MisconfigurationException, match="There should not be multiple apps instantiated within a file"):
         load_app_from_file(os.path.join(test_script_dir, "two_apps.py"))
