+++ tests/tests_examples_app/conftest.py
@@ -1,17 +1,34 @@
 import os
 import shutil
 import threading
+from subprocess import Popen
 
 import psutil
 import pytest
 
+from lightning_app import _PROJECT_ROOT
 from lightning_app.storage.path import _storage_root_dir
 from lightning_app.utilities.component import _set_context
 from lightning_app.utilities.packaging import cloud_compute
 from lightning_app.utilities.packaging.app_config import _APP_CONFIG_FILENAME
 from lightning_app.utilities.state import AppState
 
+GITHUB_APP_URLS = {
+    "template_react_ui": "https://github.com/Lightning-AI/lightning-template-react.git",
+}
+
 os.environ["LIGHTNING_DISPATCHED"] = "1"
+
+
+def pytest_sessionstart(*_):
+    """Pytest hook that get called after the Session object has been created and before performing collection and
+    entering the run test loop."""
+    for name, url in GITHUB_APP_URLS.items():
+        if not os.path.exists(os.path.join(_PROJECT_ROOT, "examples", name)):
+            path_examples = os.path.join(_PROJECT_ROOT, "examples")
+            Popen(["git", "clone", url, name], cwd=path_examples).wait(timeout=90)
+        else:
+            Popen(["git", "pull", "main"], cwd=os.path.join(_PROJECT_ROOT, "examples", name)).wait(timeout=90)
 
 
 def pytest_sessionfinish(session, exitstatus):
